ix timezone to , add igecimal for price/amount/quantity, hide null fields

+++ src/main/java/com/example/minibilling/model/domain/Invoice.java
@@ -1,5 +1,8 @@
 package com.example.minibilling.model.domain;
 
+import com.fasterxml.jackson.annotation.JsonFormat;
+
+import java.math.BigDecimal;
 import java.time.OffsetDateTime;
 import java.util.List;
 
@@ -8,8 +11,8 @@ public record Invoice(
         String documentNumber,
         String consumer,
         String reference,
-        double totalAmount,
-        double totalAmountWithVat,
+        BigDecimal totalAmount,
+        BigDecimal totalAmountWithVat,
         List<InvoiceLine> lines,
         List<VatLine> vat
 ) {}
